requirements.txt
Flask
numpy
pillow
inference_sdk
python-dotenv

